mypy compatiability with remaining files
