+++ lightworks/emulator/components/detector.py
@@ -112,10 +112,10 @@ class Detector:
         if self.efficiency == 1 and self.p_dark == 0 and self.photon_counting:
             return in_state
         # Convert state to list
-        output = [i for i in in_state]
+        output = [i for i in in_state] # type: ignore
         # Account for efficiency
         if self.efficiency < 1:
-            for mode, n in enumerate(in_state):
+            for mode, n in enumerate(in_state): # type: ignore
                 for i in range(n):
                     if random() > self.efficiency:
                         output[mode] -= 1
