+++ lightworks/emulator/results/simulation_result.py
@@ -184,7 +184,7 @@ class SimulationResult:
             for out_state, val in self.dictionary[in_state].items():
                 new_s = State([1 if s>=1 else 0 for s in out_state])
                 if invert:
-                    new_s = State([1-s for s in new_s])
+                    new_s = State([1-s for s in new_s]) # type: ignore
                 if new_s in mapped_result[in_state]:
                     mapped_result[in_state][new_s] += val
                 else:
